groq-builder: feature(compat): extracted `projectField` to separate file
